[#27681] docdb: Do not read ysql lease ttl flag when changing lease liveness.
Summary:
A background task in the master leader determines whether the YSQL lease held by tservers is live or not. This task reads the master_ysql_operation_lease_ttl_ms flag to determine this. This means any changes to this flag are immediately effective for master leaders. On the other hand tservers will only be aware of changes to this flag when they next make a successful YsqlLeaseRefresh RPC. Thus reducing this flag is not entirely safe because the master leader will immediately treat tservers as having a shorter lease TTL while tservers will not be aware of their new lease ttl until the next time they process a `YsqlLeaseRefresh` response.
Jira: DB-17395

Test Plan: n/a

Reviewers: amitanand

Reviewed By: amitanand

Subscribers: rthallam, slingam, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D44925