[BACKPORT 2025.1.0][#3109][#27650] YSQL: Reset ddl_state for read-only transactions in txn ddl
Summary:
As part of bbaf5f829069/D44196, we ensured that we set the DDL state on all intermediate YB transactions created as part of the ANALYZE command. The DDL state
is also set on the last transaction created to match the CommitTransaction of PostgresMain. This transaction doesn't do any catalog writes and is just present to match
the expectation that we will be in a transaction after we exit the `vacuum` function. Since, the last transaction doesn't do any catalog writes, it is a read-only
transaction. We don't cleanup the ddl_state store in PgTxnManager when committing a read-only transaction. This leads to the DDL state hanging around after the
completion of the transaction as well and leads to unexpected state when executing the next command.

Specifically, the following example fails:
```
create table test (a int);
create table test2(a int, b text);
analyze test, test2; -- the ddl_state in PgTxnManager is left even after the command finishes execution.
create index idx on test2(b); -- Leads to ERROR:  EnterSeparateDdlTxnMode called when already in DDL mode
```

This revision fixes the issue by adding cleanup of ddl_state even in the case of read-only transactions. Moved the logic to ResetTxnAndSession to reduce code duplication.

Also fixed an incorrect usage of the GFlag by replacing `YBCGetGFlags()` with `*YBCGetGFlags()`.
Jira: DB-1839, DB-17238

Original commit: 129ec953b515284e1e364d8104da8fd6c83bac18 / D44795

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestDdlTransactionBlocks'

Reviewers: pjain, patnaik.balivada, myang

Reviewed By: myang

Subscribers: yql, jason, myang

Differential Revision: https://phorge.dev.yugabyte.com/D44937