[#27449, #27590] DocDB: Export committed DocDB tables as of snapshot hybrid time
Summary:
Prior to this diff, export_snapshot operation was escaping some DocDB tables that were not committed on pg side. This was required to deal with cases where the table is dropped by YSQL but not docdb. Such orphaned DocDB tables are ignored while exporting the snapshot.

However, the check for pg_class status happens during the export snapshot timestamp "Let's call it t2". Whereas the snapshot is taken as of an earlier point in time "Let's call it t1" So the following scenario can happen:
1- CREATE TABLE t (k INT, v INT); ---> This creates a DocDB table D1
2- ALTER TABLE t ADD PRIMARY KEY (k);
This alter is a table rewrite alter. So a new DocDB table D2 is created that has the new uncommitted schema. Suppose that the rewrite process is a long-running one, and it is still ongoing through the following steps.
3- Create a database snapshot that includes both D1 and D2 DocDB tables. The snapshot creation time is t1 and is called `snapshot_hybrid_time`
4- Run ysql_dump as of `snapshot_hybrid_time`. So the committed DocDB table is D1.
5- Commit the Alter Table started at step 2. Therefore, DocDB table D2 is now the committed table, and D1 should be deleted.
6- Export the snapshot created at 3. In this case, DocDB table D1 and its tablets will be skipped as D1 is considered uncommitted.

We can see that the backup is inconsistent, as the ysql_dump considers DocDB table D1 to be the committed table as of t1. While the export snapshot has skipped including table D1 in its Snapshot metadata, as it read the pg catalog as of timestamp t2.

The diff fixes the issue by exporting the committed DocDB tables as of `snapshot_hybrid_time`. This means both ysql_dump and export_snapshot outputs are taken as of a consistent timestamp.

The check for the committed DocDB tables happens as follows:
Suppose `pg_table_oid` is the field that comes from TableInfo of DocDB table, `relfilenode_oid` is the suffix of the DocDB table.
Suppose the output of this query is called `committed_relfilenode`:
```SELECT relfilenode from pg_class where oid=pg_table_oid```
If `committed_relfilenode == relfilenode_oid` then this DocDB table is the committed DocDB table; otherwise, it is uncommitted.

**Upgrade/Rollback safety:**
Added a new master error type `DOCDB_TABLE_NOT_COMMITTED`, which is returned when getting the pg table oid of an uncommitted DocDB table. This error is only used in the master process.
Jira: DB-16987, DB-17166

Test Plan:
Test that only the new committed DocDB table is included in the snapshot, while the old uncommitted one is skipped:
./yb_build.sh release --cxx-test yb-admin_client-test --gtest-filter ClusterAdminClientTest.ExportCommittedDocDbTablesAsOfSnapshotHt

Test that only the old committed DocDB table is included in the snapshot, while the new uncommitted one is skipped:
./yb_build.sh release --cxx-test yb-admin_client-test --gtest-filter ClusterAdminClientTest.ExportSnapshotWithDdlInProgressTables

Reviewers: asrivastava, zdrudi, fizaa

Reviewed By: asrivastava, zdrudi, fizaa

Subscribers: ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D44897