[#27521] DocDb: Table locks need not check the flag on master
Summary:
enable_ysql_operation_lease/enable_object_locking_for_table_locks flag checks on the master.

These were added when the feature was still being developed. They are not necessary
as the master is generally a passive entity responding to the requests made by the
TServer. If the flag is enabled/disabled on the TServers, master responds accordingly.
Jira: DB-17079

Test Plan: jenkins

Reviewers: zdrudi, rthallam

Reviewed By: zdrudi

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D45284