[BACKPORT 2025.1.0][PLAT-17940] 2.18 -> 2025.1 yba upgrade fails due to a migration not completing
Summary:
2.18 -> 2025.1 yba upgrade fails due to a migration not completing.
This change will not rerun the migration if it was already applied. It'll run correctly only when we run it the first time now, which is on YBA version < `2.20.0`.
It won't rerun the migration since the checksum in this case will not change.
It also needs a better permanent fix where we point the migration model to the existing primary model itself. This should take care of the java RBAC migrations better.

Original commit: c2c066a74a50a3b33f13ecfebf95f3c618ce6056 / D45009

Test Plan:
Added a UT and checked that it fails before the fix and succeeds after the fix.
Also ran a manual test to upgrade from YBA `2.18.9-b17` to `2025.1.0.0-b<dev-itest-build>` which has this fix and the migrations ran successfully and YBA is up.
Run itests.
Run UTs.

Reviewers: amalyshev, dshubin, sneelakantan, anijhawan

Reviewed By: amalyshev

Differential Revision: https://phorge.dev.yugabyte.com/D45012