[#27630] YSQL: Raise warning instead of error when setting yb_disable_auto_analyze on a PG backend
Summary:
Before this change, setting the `yb_disable_auto_analyze` GUC via a source different from `ALTER DATABASE SET` was configured to throw an error. This resulted in failures when using connection manager along with backup/ restore (which sets the GUC on the database being restored). This happens because: PG backends that start after the `alter database set` changes the GUC value to a non-default value, return a customer parameter status to the connection manager. This is stored in the logical client connection map on conn mgr to be replayed before starting a transaction by that client. However, replaying the stored value would result in an error.

This change allows setting the GUC on backends along with a warning. This is a quick fix in order to run YbBackup* tests successfully with conn mgr which fail due to same above reason - yb_disable_auto_analyze being set using SET stmt in deploy phase. The actual fix is being tracked by: [[ https://yugabyte.atlassian.net/browse/DB-17221 | DB-17221 ]]
Jira: DB-17220

Test Plan: Make sure all ybBackUpTests are passing with connection manager

Reviewers: hsunder, pjain, skumar, rbarigidad, vikram.damle, arpit.saxena

Reviewed By: pjain

Subscribers: yql, hsunder, arpit.saxena

Differential Revision: https://phorge.dev.yugabyte.com/D44710