[PLAT-17877] Writing prometheus logs to data directory
Summary: Prometheus logs for YBA in yba-installer were being written to /var/log in case of issues with SELinux and systemd services logging to custom directories like /opt/yugabyte. The problem doesn't seem to exist anymore and was initially added as part of D38761.

Test Plan:
Install on ubuntu 20, 22, centos 7 with SELinux enforcing
Install on almalinux 9 - this is what was causing problems with /opt/yugabyte/data/logs initially and necessitated /var/log

Reviewers: dshubin

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D44972