[#23199] CDC, ASH: YSQL side instrumentation
Summary:
This diff adds instrumentation for RPCs sent to
CDC Service from PG client and sends the
metadata as part of the RPCs.

**Upgrade/Downgrade safety:**
The new fields are optional, the data for ASH will
not be available temporarily and that is okay.
Jira: DB-12142

Test Plan: ./yb_build.sh --cxx-test pg_ash-test --gtest_filter PgWaitEventAuxTest.PgCDCServiceRPCs

Reviewers: dmitry, sumukh.phalgaonkar

Reviewed By: dmitry, sumukh.phalgaonkar

Subscribers: hbhanawat, yql

Differential Revision: https://phorge.dev.yugabyte.com/D45048