[#27644] YSQL: Wait for stats update before dropping database with Connection Manager
Summary:
DROP DATABASE operations are dependent on the logical connection count when using Connection Manager, rather than the physical connection count. Freshness of logical connection stats is dependent on reading from shared memory, which is further dependent on ysql connection manager stats refresh interval. This diff introduces an internal sleep for the duration of the stats refresh duration so as to read an appropriate logical connection count for operations such as:
```
\c custom_database
-- perform operations...
\c yugabyte
DROP DATABASE yugabyte;
```

This order of queries usually faces an error like "1 session is still using the database "custom_database" ", but with the help of the sleep before dropping the database, such race conditions would no longer be an issue.

Jira: DB-17231

Test Plan: Jenkins: all tests

Reviewers: skumar, mkumar, arpit.saxena, vikram.damle, asrinivasan

Reviewed By: skumar

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D44753