[#27695] YSQL: Set range scan bounds at PgDocReadOp initialization
Summary:
There is an existing function to calculate scan bounds out of the request's
conditions. Previously this function was used in two distinct places,
independently for forward and backward scans. Calculated bounds weren't
stored with the request and therefore were recalculated for every page.
There was code paths where bounds weren't calculated at all, causing
avoidable requests to partitions without matches.
This diff addresses those issues by calculating bounds earlier in execution
process, and storing values in respective request's fields, so they can be
referenced by both PgGate and DocDB without need to recalculate.
Now bounds are calculated when PgDocReadOp is initialized. By that time
the template request is finalized, that means the parameters on which the
request bounds depend won't change. So the bounds can be calculated
and combined with already existing bounds. If multiple bounds are combined,
the resulting range can be empty. That's the case when PgGate can skip
sending the requests and save resources.
That eliminates necessity to recalculate bounds later, as bounds are
copied when we clone tablet requests.
Jira: DB-17294

Test Plan:
./yb_build.sh --java-test 'org.yb.pgsql.TestPgSequentialSelect#testRangeIndexScanOnHashTable'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgSequentialSelect#testRangeIndexScanOnAscTable'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgSequentialSelect#testRangeIndexScanOnDescTable'

Reviewers: arybochkin, dmitry, jason, sergei, timur

Reviewed By: jason

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D44893