[PLAT-17963] Identify whether to use use_sudo to run ansible runbooks from a more deterministic way
Summary:
This diff updates the way that YBA determines whether the otel-collector service is running at user scope or system scope by checking the existence of the tserver systemd unit file.

Previously, it was relying whether the tserver service status can be checked with `--user` argument and if not, it used system scope. However, on the universes with dedicated master nodes, the tserver is not running and this logic would cause the otel-collector to be managed in the system scope even when it is user scope.

Test Plan: Created a universe with dedicated master nodes and enabled audit logging to GCP and then ran a gflag upgrade using rolling restart which was successful

Reviewers: anijhawan, svarshney, nsingh, amalyshev

Reviewed By: anijhawan, svarshney

Subscribers: skurapati, anijhawan

Differential Revision: https://phorge.dev.yugabyte.com/D45092