[PLAT-17856]: Checkbox in kubernetes overrides upgrade modal is getting overlapped
Summary:
Checkbox in kubernetes overrides upgrade modal is getting overlapped

One of the issues in YBModalForm is that, for the entire footerAccessory, it assigns pull-left or pull-right, this does not allow is in the footer to individually allow elements to float left or right. If I change this behavior in YBModalForm, it needs extensive testing on all other dialogs as well.
However newer components dont use YBModalForm as it is a legacy component,

Test Plan:
Please refer to screenshot

BEFORE
{F371238}

AFTER
{F371237}

Reviewers: jmak

Reviewed By: jmak

Differential Revision: https://phorge.dev.yugabyte.com/D45256