[PLAT-18058] Changing port in CP is failing due to manual setting of pgsql_proxy_bind_address
Summary:
Changing port in CP is failing due to manual setting of pgsql_proxy_bind_address.
This is an add-on to the fixes done in (https://phorge.dev.yugabyte.com/D42741) and (https://phorge.dev.yugabyte.com/D42741).

We need to take up the taskParams overridden port values when we are trying to change the ports in the corner case where the user is manually setting the gflag `pgsql_proxy_bind_address`. YBM also currently does it this way. This diff will solidify the fix.

Test Plan:
Manually tested the following flow before and after this fix:
Before:
1. Created a universe with CP off, ran configureYSQL with setting `internalYsqlServerRpcPort=6200`.
2. Enabled CP via configureYSQL API and set the `internalYsqlServerRpcPort=6433`.
3. Observed this failed at the WaitForServer step where the PG process was actually started on 6200 instead of the intended 6433.

After:
1. Created a universe with CP off, ran configureYSQL with setting `internalYsqlServerRpcPort=6200`.
2. Enabled CP via configureYSQL API and set the `internalYsqlServerRpcPort=6433`.
3. Observed this succeeded and the PG process started at 6433 as intended.

Also run UTs.
Ran itests.

Reviewers: sneelakantan

Reviewed By: sneelakantan

Differential Revision: https://phorge.dev.yugabyte.com/D45267