[BACKPORT 2024.2] [#22472, #22473] YSQL: Support hypopg in old and new cost models
Summary:
Original commit: 6f8f81f9a6dfe433345966a7d840a86a33fba5ae / D44042
Hypopg is a PostgreSQL extension that allows users to quickly create
hypothetical indexes before actually creating real indexes to verify that the
query planner would pick these indexes.

The hypothetical indexes are only made visible to the planner when the query is
run with `EXPLAIN`. This is done by adding `IndexOptInfo` objects for the
hypothetical indexes to the list of indexes for each table. The planner only
relies on these objects for most part and does not realize that the indexes do
not exist in reality.

Before this change, the old cost model logic had been bypassed for hypothetical
indexes and the cost estimates were not accurate. This limits the utility of
hypothetical indexes. In fact, some variables were left uninitialized which led
to intermittent crashes and test instability (#22473). The new cost model
(`yb_enable_cbo=ON`) would crash when hypothetical indexes were used.

In this change, the old cost model code has been refactored to allow correctly
estimating the cost of the hypothetical indexes. Specifically, the old cost
model would try to fetch the `Index` object for the hypothetical index which
does not exist to get information about the index keys. This information is now
gathered from `IndexOptInfo` instead.

The crash in the new cost model, was also caused due to a similar reason and
some code has been refactored in this change.

With these changes, the cost estimate for the hypothetical index should now
match the cost estimate for an actual index with the same definition.
Jira: DB-11388, DB-11389

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressThirdPartyExtensionsHypopg'

Reviewers: mihnea, william.mckenna

Reviewed By: william.mckenna

Subscribers: jason, yql

Differential Revision: https://phorge.dev.yugabyte.com/D45280