[#27923] docdb: Check prefetched key is within bounds before passing into GetHelper
Summary:
`GetHelper<>::Run()` takes in the result of the first `Fetch()` call on the underlying
iterator (or in some cases does the first `Fetch()` call itself). Additionally, `GetHelper` sets the upper
bound on the iterator and iterates through the rest of the entries.

But we do not properly handle the case where the first `Fetch()` call is already out of bounds.
We only have DCHECKs for ensuring entries are in bounds, and so in release builds, we strip away the
root doc key based solely on length (even though it may not be a prefix) and then attempt to decode
the rest of the key. This leads to bad behavior, especially in the case where length of the key is
less than the length of the root doc key:
- The sub key slice now has begin > end, so calling size() on the slice returns a large number
  (since size is calculated as `static_cast<size_t>(end - begin)`).
- If the first character of the sub key slice (positioned at some garbage byte) is unknown, we
  attempt to format it with FormatBytesAsStr for the error message.
- FormatBytesAsStr takes start pointer + size, so we pass in a large (>2^63) value for size.
- This can access invalid addresses, causing SIGSEGV.
- In versions without the fix from D45331, this also ignores the max_length field, causing us
  to copy from memory continuously until we run out of memory or hit invalid addresses and
  get a SIGSEGV.

This change adds a check that the root doc key is indeed a prefix of the prefetched key before
continuing in this case.
Jira: DB-17539

Test Plan: Jenkins

Reviewers: sergei, arybochkin

Reviewed By: sergei

Subscribers: rthallam, arybochkin, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D45293