[PLAT-18063] API to get YBA FIPS status
Summary:
We need to show if YBA instance is FIPS compliant or not in the UI.
Hence, adding the API for that.
It will be used to get any other static configuration/info we need to expose later as well.

Request:
```
GET /api/v2/yba-info
```
Response:
```
{
	"fips_enabled": false
}
```

Test Plan: Tested API manually.

Reviewers: dshubin

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D45295