[#27927] YSQL: Fix crash in using planner hints with hypothetical indexes
Summary:
Using planner hints to force Index Scan with a particular index on a table
which has hypothetical indexes, would lead to a crash.

```
CREATE TABLE test (v1 INT, v2 INT);
INSERT INTO test (SELECT s, s FROM generate_series(1, 10000) s);
ANALYZE test;
SELECT hypopg_create_index('CREATE INDEX ON test (v1 ASC)');
CREATE INDEX test_real_idx ON test (v1 ASC);
EXPLAIN /*+ IndexScan(test test_real_idx) */ SELECT * FROM test WHERE v1 = 1;
server closed the connection unexpectedly
        This probably means the server terminated abnormally
        before or while processing the request.
The connection to the server was lost. Attempting reset: Succeeded.
```

This was caused by the `pg_hint_plan` extension trying to fetch the name of the
hypothetical index. Hypothetical indexes are only visible to the planner when
a query is run using `EXPLAIN`. The `hypopg` extension adds `IndexOptInfo`
objects for the hypothetical indexes, but the index does not really exist.
The `IndexOptInfo` object does not contain the name of the index, so when
the `pg_hint_plan` extension tries to fetch the name of the index to enforce
the hints, it tries to access the corresponding `Index` object, which does not
exist.

To fix this, we reuse a callback registered by the `hypopg` index, that returns
the name of the hypothetical index. This resolves the crash and also allows
us to force the use of the hypothetical index using hints.
Jira: DB-17543

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressThirdPartyExtensionsHypopg'

Reviewers: mihnea, william.mckenna

Reviewed By: william.mckenna

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D45303