[#27945] YSQL: ALTER ROLE ... SET GUC needs to have global-impact
Summary:
Currently, if a DDL statement writes to a shared relation that also has a
catalog cache, we mark this DDL statement to have global-impact. However ALTER
ROLE ... SET GUC writes to the shared relation `pg_db_role_setting` that does
not have a catalog cache. As a result it is not marked as having global-impact.

This has problem when tserver response cache is enabled. Assume the DDL is
executed from a connection to DB1. Because this DDL only increments the catalog
version of DB1, other DBs will not have their catalog version incremented. As a
result, a new connection to DB2 will continue to load the stale contents of
`pg_db_role_setting` of DB2 from tserver response cache, and the effect of the
DDL is not seen.

To fix this bug, I added a new function `YbSharedRelationIdNeedsGlobalImpact`
that adds shared relations that need to have global impact and added
`pg_db_role_setting` there. If similar bugs are found for other shared
relations, we can add more there.

A new unit test is added that would fail before this diff.
Jira: DB-17560

Test Plan: ./yb_build.sh release --cxx-test pg_catalog_version-test --gtest_filter PgCatalogVersionTest.TestAlterRoleSetGUCHasGlobalImpact

Reviewers: kfranz, sanketh, mihnea

Reviewed By: sanketh

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D45329