[PLAT-18093] Extract azcopy/gutil in case they are present
Summary:
We recently started packaging the Node Agent on architectures basis. As part of this effort, we now filter out x86-based dependencies when building for ARM.
Previously, we were installing the AMD64 version of azcopy on ARM machines, which was not appropriate.
Since these utilities are no longer used on the master for backup, it is safe to conditionally extract them only if they are present.

Test Plan: Manually created aarm based universe.

Reviewers: vkumar, anijhawan, nsingh

Reviewed By: anijhawan

Differential Revision: https://phorge.dev.yugabyte.com/D45358