[#27929] Ysql: Add V84__26540__yb_int_pg_stats_v11 migration script
Summary:
When performing upgrades we ensure the target version is newer by checking the list of AutoFlags and ysql_migration scripts between the two versions. Removal of AutoFlags or ysql_migration scripts is not allowed.
`V59.7__26540__yb_int_pg_stats_v11.sql` was added to 2024.2 and not in master since it added the `yb_int_pg_stats_v11` view which was only required in pg11, and not in pg15. Since the script does not exist in master/2025.1 it cases upgrades to fail. Adding a new empty script to unblock upgrades.

Updated `PGConn::Execute` to support empty queries, or queries that only contain comments.
Jira: DB-17545

Test Plan: ybd --cxx-test basic_upgrade-test --gtest_filter UpgradeFrom_2024_2_3_0/BasicUpgradeTest.TestUpgrade/0

Reviewers: fizaa

Reviewed By: fizaa

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D45372