[#27986] docdb: Fix error attempting to decode post-apply metadata as reverse index
Summary:
559b2b06244 adds a post-apply transaction metadata record into intentsdb when CDC
is enabled with long retention changes. The post-apply metadata has a key of `[txn id] 00`,
so we end up with:
```
- [txn id] - txn metadata
- [txn id] 00 - txn post apply metadata (only if CDC long retention enabled)
- [txn id] [hybrid time1] - reverse index entry
- [txn id] [hybrid time2] - reverse index entry
- ...
```

`TransactionLoader` scans backwards from the last reverse index entry, and parses reverse index
entries. But it is currently set to terminate the loop when it finds a key that is not a strict
prefix of `[txn id]`, which means that it attempts to parse the post apply metadata as a reverse
index entry. This always fails (`00` is not a hybrid time), and we simply log an error and skip
the record, stopping in the next iteration (txn metadata):
```
E0715 10:00:00.000000 12345 transaction_loader.cc:316] T 0123456789abcdef0123456789abcdef P 0123456789abcdef0123456789abcdef: Failed to decode intent while loading transaction 01234567-89ab-cdef-0123-456789abcdef, 780123456789ABCDEF0123456789ABCDEF00 => 0A100123456789ABCDEF0123456789ABCDEF1201020304FEFE0119FEFFFFFFFFFFFFFF21FEFFFFFFFFFFFFFF: Corruption (yb/common/doc_hybrid_time.cc:227): Encoded HybridTime can't be more than 30 bytes, found 31: 31 vs 30
```

This changes the logic to stop the loop when we find a post apply metadata record, so that the
(harmless) error is not emitted.
Jira: DB-17601

Test Plan: Jenkins

Reviewers: sergei, huapeng.yuan

Reviewed By: huapeng.yuan

Subscribers: huapeng.yuan, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D45405